app/app/urls.py
from django.conf import settings
from django.conf.urls.static import static
from django.contrib import admin
from django.urls import path, include


urlpatterns = [
    path("", include("core.urls")),
    path("admin/", admin.site.urls),
]


if settings.DEBUG:  # pragma: no cover
    import debug_toolbar

    urlpatterns = (
        urlpatterns
        + [path("__debug__/", include(debug_toolbar.urls))]  # noqa: W503
        + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)  # noqa: W503
    )
