requirements.txt
Django==4.2.1
django-crispy-forms==2.0
crispy-bootstrap5==0.7
psycopg2==2.9.6
gunicorn==20.1.0
django-debug-toolbar
pre-commit
sentry-sdk
coverage
